Refactor TTS and input manager into modules
